requirements.txt
selenium==4.32.0
Appium-Python-Client==5.1.1
pytest==8.3.5
pyyaml==6.0.2
pandas==2.2.3
openpyxl==3.1.5
faker==37.1.0
colorlog==6.9.0